+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/di/WearAppModule.kt
@@ -2,6 +2,9 @@ package au.com.shiftyjelly.pocketcasts.wear.di
 
 import android.content.Context
 import android.net.ConnectivityManager
+import au.com.shiftyjelly.pocketcasts.NoOpGravatarSdkService
+import au.com.shiftyjelly.pocketcasts.utils.gravatar.GravatarService
+import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -10,11 +13,16 @@ import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(SingletonComponent::class)
-object WearAppModule {
+abstract class WearAppModule {
 
-    @Provides
-    fun connectivityManager(
-        @ApplicationContext application: Context,
-    ): ConnectivityManager =
-        application.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+    companion object {
+        @Provides
+        fun connectivityManager(
+            @ApplicationContext application: Context,
+        ): ConnectivityManager =
+            application.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+    }
+
+    @Binds
+    abstract fun gravatarService(factory: NoOpGravatarSdkService.Factory): GravatarService.Factory
 }